WIP: Fix session-per-service refactor issues
- Fixed transaction management in ProcessorHandler for v2 processors
- Modified ProcessingService to accept optional db_manager parameter
- Removed redundant commit in ProcessingService._record_state_transition()
- Added debugging logs to track thread safety issues with API token usage
- Identified root cause: SQLAlchemy sessions shared across threads in Azure Functions

TODO: Implement session-per-execution pattern to ensure thread safety

+++ src/db/db_state_transition_models.py
@@ -7,7 +7,7 @@ with a focus on tenant isolation and clean architecture.
 
 import enum
 
-from sqlalchemy import Column, ForeignKey, Index, Integer, String, Text
+from sqlalchemy import BigInteger, Column, ForeignKey, Index, Integer, String, Text
 from sqlalchemy.orm import relationship
 from sqlalchemy.types import Enum
 
@@ -62,7 +62,7 @@ class StateTransition(Base, BaseModel):
         Integer, nullable=True, comment="Time (in ms) spent in the previous state"
     )
     sequence_number = Column(
-        Integer, nullable=True, comment="Sequence number for ordering transitions"
+        BigInteger, nullable=True, comment="Sequence number for ordering transitions"
     )
 
     # Explicit tenant association with cascading delete
